Refactor application modules into CQRS structure
